se realizao mejoras en todos las apirequest, se introdujo la opcion de response entity

+++ usuarios/src/main/java/com/citas/usuarios/repository/PrecioRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.citas.usuarios.entity.Precios;
 
-public interface PrecioRepository extends JpaRepository<Precios, Integer>{
+public interface PrecioRepository extends JpaRepository<Precios, Long>{
     //hacer una querry con la factura y los precios
 
-   // Precios findById(Integer idPrecio);
+
 
 }
